+++ CHANGELOG.md
@@ -19,3 +19,4 @@
 01. :gift: Added class `dev.orne.http.client.AbstractStatusDependentOperation`
 01. :gift: Added class `dev.orne.http.client.VoidResponseHandler`
 01. :gift: Added class `dev.orne.http.client.XMLResponseHandler`
+01. :gift: Added class `dev.orne.http.client.JSONResponseHandler`